// packages/cli-server/generated/cli.schemas.ts
export type CliOrgCapability = typeof CliOrgCapability[keyof typeof CliOrgCapability];


export const CliOrgCapability = {
  orglist: 'org.list',
  orgread: 'org.read',
  sourceconnect: 'source.connect',
  sourcelist: 'source.list',
  sourceread: 'source.read',
  queryexecute: 'query.execute',
} as const;

export interface CliOrgReadResponse {
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  name?: string;
  roles?: string[];
  capabilities?: CliOrgCapability[];
}