// packages/cli-server/generated/cli.zod.ts
export const cliOrgGetPathOrgSlugRegExp = new RegExp('^[a-z0-9]+(?:-[a-z0-9]+)\*$');


export const CliOrgGetParams = zod.object({
  "orgSlug": zod.string().min(1).regex(cliOrgGetPathOrgSlugRegExp)
})




export const CliOrgGetQueryParams = zod.object({
  "fields": zod.string().min(1).optional()
})








export const CliOrgGetResponse = zod.object({
  "requestId": zod.string().min(1),
  "data": zod.object({
  "slug": zod.string().min(1).optional(),
  "name": zod.string().min(1).optional(),
  "roles": zod.array(zod.string().min(1)).optional(),
  "capabilities": zod.array(zod.enum(['org.list', 'org.read', 'source.connect', 'source.list', 'source.read', 'query.execute'])).optional()
}),
  "warnings": zod.array(zod.string().min(1))
})